Add transport discovery visualizer CLI command and shared package
- Add pkg/tpviz package with transport discovery visualizer server
- Add 'skywire cli tp viz' command to start standalone visualizer
- Update rewards server to use shared tpviz package
- Remove duplicate transportGraphHTML from rewards server
